\name{mix.print}
\alias{mix.print}
\title{Printing mix object}
\description{Printing a \code{smsn.mix} object (univariate case, p=1)}
\usage{
mix.print(model, digits = 3, ...)
}
\arguments{
  \item{model}{an object of class \code{snsm.mix}, see \code{\link{smsn.mix}} for details}
  \item{digits}{rounding for tabular output on the console
    (default is to round to 3 decimal place)}
  \item{...}{further arguments to \code{\link{print}}}
}
\seealso{\code{\link{smsn.mix}}}
\keyword{print}
